Refactorizacion de código.
